Half-complete Village Scripts
